Rename snakeserver/ to benchmark/
Consolidates all bench tooling under the benchmark/ directory. The
snakeserver bench_suite.go (full-featured: parallel execution, -easing
flag, -filter flag, sine/constant/linear profiles) replaces the simpler
version that was in benchmark/. Also moves the HTTP server (main.go),
go module, and plot_suite.py.

Updates plot_easing_comparison.py --bench-go default path to reference
the local bench_suite.go now that it lives in the same directory.

Removes the accidentally-tracked snakeserver binary.

AI disclosure: Claude Code assisted with development. Every line of code was either written by or carefully reviewed by me :)
