+++ go/vt/vttablet/tabletserver/loadshed/benchmark/bench_suite.go
@@ -4,21 +4,23 @@ package main
 
 import (
 	"context"
+	"encoding/csv"
 	"flag"
 	"fmt"
 	"math"
 	"os"
-	"path/filepath"
+	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"vitess.io/vitess/go/vt/vttablet/tabletserver/loadshed"
 )
 
+var easingDiv *float64
+
 type event struct {
 	tsMs      int64
-	eventType string
+	kind      string // "issued", "granted", "shed"
 	latencyMs float64
 }
 
@@ -29,84 +31,93 @@ type statsSnapshot struct {
 	holderCount     int
 	dropping        bool
 	dropCount       int
-	currentInterval int64
+	currentInterval int64 // ns
+	lastDropsPerRun int
 }
 
-type loadProfile func(elapsed, total time.Duration) float64
+// loadProfile returns the load fraction [0,1] at a given elapsed time.
+type loadProfile func(elapsed, totalDuration time.Duration) float64
 
-func linearRamp(elapsed, total time.Duration) float64 {
-	return elapsed.Seconds() / total.Seconds()
+func sineProfile(period time.Duration) loadProfile {
+	return func(elapsed, _ time.Duration) float64 {
+		phase := float64(elapsed) / float64(period) * 2 * math.Pi
+		return (1 - math.Cos(phase)) / 2
+	}
 }
 
-func runTest(cfg testConfig) ([]event, []statsSnapshot) {
-	capacity := cfg.capacity
-	workMs := cfg.workMs
-	targetMs := cfg.targetMs
-	intervalMs := cfg.intervalMs
-	exponent := cfg.exponent
-	totalDuration := time.Duration(cfg.durationMs) * time.Millisecond
-	peakArrivalRateMultiplier := cfg.peakArrivalRateMultiplier
-	profile := cfg.profile
+func constantProfile() loadProfile {
+	return func(_, _ time.Duration) float64 {
+		return 1.0
+	}
+}
 
+func linearRampProfile() loadProfile {
+	return func(elapsed, totalDuration time.Duration) float64 {
+		return float64(elapsed) / float64(totalDuration)
+	}
+}
+
+func runBench(capacity int, peakArrivalRateMultiplier float64, durationMs, workMs, targetMs, intervalMs int, profile loadProfile) ([]event, []statsSnapshot) {
 	snake := loadshed.NewSnake(loadshed.SnakeConfig{
-		Name: "bench",
+		Name:     "bench",
+		Capacity: func() int { return capacity },
+		MaxAge:   func() time.Duration { return 30 * time.Second },
 		CoDel: loadshed.CoDelConfig{
 			TargetNs:       func() int64 { return int64(targetMs) * 1_000_000 },
 			IntervalNs:     func() int64 { return int64(intervalMs) * 1_000_000 },
 			MinDropDelayNs: func() int64 { return 1_000_000 },
-			Exponent:       func() float64 { return exponent },
-			EasingDivisor:  func() float64 { return 1.2 },
+			Exponent:       func() float64 { return 1 },
+			EasingDivisor:  func() float64 { return *easingDiv },
 		},
-		Capacity:            func() int { return capacity },
-		MaxAge:              func() time.Duration { return 30 * time.Second },
 		LoadsheddingAllowed: func() bool { return true },
 	})
 
-	ctx, cancel := context.WithCancel(context.Background())
+	totalDuration := time.Duration(durationMs) * time.Millisecond
+	ctx, cancel := context.WithTimeout(context.Background(), totalDuration+500*time.Millisecond)
+	defer cancel()
+
+	start := time.Now()
 
 	var mu sync.Mutex
 	var events []event
-	var stats []statsSnapshot
-	start := time.Now()
 
-	record := func(evtType string, latMs float64) {
+	record := func(kind string, latencyMs float64) {
 		ts := time.Since(start).Milliseconds()
 		mu.Lock()
-		events = append(events, event{tsMs: ts, eventType: evtType, latencyMs: latMs})
+		events = append(events, event{tsMs: ts, kind: kind, latencyMs: latencyMs})
 		mu.Unlock()
 	}
 
-	// Stats polling goroutine
+	// Stats sampling goroutine — sample every 5ms
+	var stats []statsSnapshot
 	statsDone := make(chan struct{})
 	go func() {
 		defer close(statsDone)
-		ticker := time.NewTicker(10 * time.Millisecond)
-		defer ticker.Stop()
+		statsTicker := time.NewTicker(5 * time.Millisecond)
+		defer statsTicker.Stop()
 		for {
 			select {
 			case <-ctx.Done():
 				return
-			case <-ticker.C:
+			case <-statsTicker.C:
 				s := snake.Stats()
-				ts := time.Since(start).Milliseconds()
-				mu.Lock()
 				stats = append(stats, statsSnapshot{
-					tsMs:            ts,
+					tsMs:            time.Since(start).Milliseconds(),
 					queueLen:        s.QueueLen,
 					droppableLen:    s.DroppableLen,
 					holderCount:     s.HolderCount,
 					dropping:        s.Dropping,
 					dropCount:       s.DropCount,
 					currentInterval: s.CurrentInterval,
+					lastDropsPerRun: s.LastDropsPerRun,
 				})
-				mu.Unlock()
 			}
 		}
 	}()
 
 	var wg sync.WaitGroup
-	var issued atomic.Int64
 
+	// Peak arrival rate as multiple of system throughput
 	systemThroughput := float64(capacity) * 1000.0 / float64(workMs)
 	peakArrivalRate := peakArrivalRateMultiplier * systemThroughput
 
@@ -137,7 +148,6 @@ func runTest(cfg testConfig) ([]event, []statsSnapshot) {
 
 			for range toSpawn {
 				wg.Add(1)
-				issued.Add(1)
 				go func() {
 					defer wg.Done()
 
@@ -172,95 +182,208 @@ type testConfig struct {
 	workMs                    int
 	targetMs                  int
 	intervalMs                int
-	exponent                  float64
 	profile                   loadProfile
 }
 
+type testResult struct {
+	cfg                   testConfig
+	issued, granted, shed int
+	events                []event
+	stats                 []statsSnapshot
+}
+
 func main() {
-	outDir := flag.String("out", "", "Output directory for TSVs (default: ~/snake-load-test-charts/tsv/<date>/<timestamp>/)")
-	exponent := flag.Float64("exponent", 1.0, "CoDel exponent (production default: 1.0)")
-	targetMs := flag.Int("target-ms", 5, "CoDel target in ms (production default: 5)")
-	intervalMs := flag.Int("interval-ms", 100, "CoDel interval in ms (production default: 100)")
+	outDir := flag.String("out", "", "Output directory (default: timestamped under ~/snake-load-test-charts/)")
+	parallel := flag.Int("j", 8, "Max parallel benchmarks")
+	filter := flag.String("filter", "", "Only run tests whose label contains this substring")
+	easingDiv = flag.Float64("easing", 2.0, "Easing divisor for CoDel count decay")
 	flag.Parse()
 
 	if *outDir == "" {
-		now := time.Now()
-		date := now.Format("2006-01-02")
-		ts := now.Format("2006-01-02_15-04-05")
 		home, _ := os.UserHomeDir()
-		*outDir = filepath.Join(home, "snake-load-test-charts", "tsv", date, ts)
+		*outDir = fmt.Sprintf("%s/snake-load-test-charts/%s", home, time.Now().Format("2006-01-02_15-04-05"))
 	}
-	os.MkdirAll(*outDir, 0o755)
+	os.MkdirAll(*outDir, 0755)
 
-	tests := []testConfig{
-		{
-			label:                     "linear_ramp__0_to_80x_cap__work_half_target",
-			capacity:                  10,
+	capacity := 10
+	targetMs := 5
+	intervalMs := 100
+
+	works := []struct {
+		label string
+		ms    int
+	}{
+		{"half_target", targetMs / 2},
+		{"equal_target", targetMs},
+	}
+
+	var configs []testConfig
+
+	// Sine wave tests: 3 periods × 3 peaks × 2 work durations, duration = 2 full periods
+	periods := []struct {
+		label string
+		ms    int
+	}{
+		{"2x_interval", 200},
+		{"10x_interval", 1000},
+		{"20x_interval", 2000},
+	}
+
+	peaks := []struct {
+		label      string
+		multiplier float64
+	}{
+		{"20x_cap", 20},
+		{"100x_cap", 100},
+		{"300x_cap", 300},
+	}
+
+	for _, p := range periods {
+		for _, pk := range peaks {
+			for _, w := range works {
+				label := fmt.Sprintf("sine__period_%s__peak_%s__work_%s", p.label, pk.label, w.label)
+				configs = append(configs, testConfig{
+					label:                     label,
+					capacity:                  capacity,
+					peakArrivalRateMultiplier: pk.multiplier,
+					durationMs:                p.ms * 2,
+					workMs:                    w.ms,
+					targetMs:                  targetMs,
+					intervalMs:                intervalMs,
+					profile:                   sineProfile(time.Duration(p.ms) * time.Millisecond),
+				})
+			}
+		}
+	}
+
+	// Constant load at 5x capacity, 20s
+	for _, w := range works {
+		label := fmt.Sprintf("constant__5x_cap__work_%s", w.label)
+		configs = append(configs, testConfig{
+			label:                     label,
+			capacity:                  capacity,
+			peakArrivalRateMultiplier: 5,
+			durationMs:                20000,
+			workMs:                    w.ms,
+			targetMs:                  targetMs,
+			intervalMs:                intervalMs,
+			profile:                   constantProfile(),
+		})
+	}
+
+	// Linear ramp from 0x to 80x capacity, 20s
+	for _, w := range works {
+		label := fmt.Sprintf("linear_ramp__0_to_80x_cap__work_%s", w.label)
+		configs = append(configs, testConfig{
+			label:                     label,
+			capacity:                  capacity,
 			peakArrivalRateMultiplier: 80,
 			durationMs:                20000,
-			workMs:                    2,
-			targetMs:                  *targetMs,
-			intervalMs:                *intervalMs,
-			exponent:                  *exponent,
-			profile:                   linearRamp,
-		},
+			workMs:                    w.ms,
+			targetMs:                  targetMs,
+			intervalMs:                intervalMs,
+			profile:                   linearRampProfile(),
+		})
 	}
 
-	for _, tc := range tests {
-		fmt.Printf("Running: %s (capacity=%d, work=%dms, target=%dms, interval=%dms, exponent=%.2f)\n",
-			tc.label, tc.capacity, tc.workMs, tc.targetMs, tc.intervalMs, tc.exponent)
-		events, stats := runTest(tc)
+	if *filter != "" {
+		var filtered []testConfig
+		for _, c := range configs {
+			if strings.Contains(c.label, *filter) {
+				filtered = append(filtered, c)
+			}
+		}
+		configs = filtered
+	}
 
-		// Write events TSV
-		tsvPath := filepath.Join(*outDir, tc.label+".tsv")
-		f, err := os.Create(tsvPath)
+	fmt.Printf("Running %d benchmarks (parallelism=%d)...\n", len(configs), *parallel)
+
+	results := make([]testResult, len(configs))
+	sem := make(chan struct{}, *parallel)
+	var wg sync.WaitGroup
+
+	for i, cfg := range configs {
+		wg.Add(1)
+		sem <- struct{}{}
+		go func(idx int, c testConfig) {
+			defer wg.Done()
+			defer func() { <-sem }()
+
+			fmt.Printf("  START [%d/%d] %s\n", idx+1, len(configs), c.label)
+			events, stats := runBench(c.capacity, c.peakArrivalRateMultiplier, c.durationMs, c.workMs, c.targetMs, c.intervalMs, c.profile)
+
+			var issued, granted, shed int
+			for _, ev := range events {
+				switch ev.kind {
+				case "issued":
+					issued++
+				case "granted":
+					granted++
+				case "shed":
+					shed++
+				}
+			}
+			results[idx] = testResult{cfg: c, issued: issued, granted: granted, shed: shed, events: events, stats: stats}
+			shedRate := 0.0
+			if issued > 0 {
+				shedRate = float64(shed) / float64(issued) * 100
+			}
+			fmt.Printf("  DONE  [%d/%d] %s → %d issued, %d granted, %d shed (%.1f%%)\n",
+				idx+1, len(configs), c.label, issued, granted, shed, shedRate)
+		}(i, cfg)
+	}
+
+	wg.Wait()
+
+	// Write TSV files into a tsv/ subdirectory
+	tsvDir := fmt.Sprintf("%s/tsv", *outDir)
+	os.MkdirAll(tsvDir, 0755)
+	for _, r := range results {
+		path := fmt.Sprintf("%s/%s.tsv", tsvDir, r.cfg.label)
+		f, err := os.Create(path)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "ERROR creating %s: %v\n", tsvPath, err)
+			fmt.Printf("  ERROR writing %s: %v\n", path, err)
 			continue
 		}
-		fmt.Fprintf(f, "ts_ms\tevent\tlatency_ms\n")
-		for _, e := range events {
-			latStr := ""
-			if e.eventType == "granted" {
-				latStr = fmt.Sprintf("%.3f", e.latencyMs)
-			}
-			fmt.Fprintf(f, "%d\t%s\t%s\n", e.tsMs, e.eventType, latStr)
+		csvW := csv.NewWriter(f)
+		csvW.Comma = '\t'
+		csvW.Write([]string{"ts_ms", "event", "latency_ms"})
+		for _, ev := range r.events {
+			csvW.Write([]string{fmt.Sprintf("%d", ev.tsMs), ev.kind, fmt.Sprintf("%.3f", ev.latencyMs)})
 		}
+		csvW.Flush()
 		f.Close()
 
 		// Write stats TSV
-		statsPath := filepath.Join(*outDir, tc.label+"_stats.tsv")
+		statsPath := fmt.Sprintf("%s/%s_stats.tsv", tsvDir, r.cfg.label)
 		sf, err := os.Create(statsPath)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "ERROR creating %s: %v\n", statsPath, err)
+			fmt.Printf("  ERROR writing %s: %v\n", statsPath, err)
 			continue
 		}
-		fmt.Fprintf(sf, "ts_ms\tqueue_len\tdroppable_len\tholder_count\tdropping\tdrop_count\tcurrent_interval_ns\n")
-		for _, s := range stats {
-			droppingInt := 0
+		csvS := csv.NewWriter(sf)
+		csvS.Comma = '\t'
+		csvS.Write([]string{"ts_ms", "queue_len", "droppable_len", "holder_count", "dropping", "drop_count", "current_interval_ns", "last_drops_per_run"})
+		for _, s := range r.stats {
+			dropping := "0"
 			if s.dropping {
-				droppingInt = 1
+				dropping = "1"
 			}
-			fmt.Fprintf(sf, "%d\t%d\t%d\t%d\t%d\t%d\t%d\n",
-				s.tsMs, s.queueLen, s.droppableLen, s.holderCount, droppingInt, s.dropCount, s.currentInterval)
+			csvS.Write([]string{
+				fmt.Sprintf("%d", s.tsMs),
+				fmt.Sprintf("%d", s.queueLen),
+				fmt.Sprintf("%d", s.droppableLen),
+				fmt.Sprintf("%d", s.holderCount),
+				dropping,
+				fmt.Sprintf("%d", s.dropCount),
+				fmt.Sprintf("%d", s.currentInterval),
+				fmt.Sprintf("%d", s.lastDropsPerRun),
+			})
 		}
+		csvS.Flush()
 		sf.Close()
-
-		var issuedCount, grantedCount, shedCount int
-		for _, e := range events {
-			switch e.eventType {
-			case "issued":
-				issuedCount++
-			case "granted":
-				grantedCount++
-			case "shed":
-				shedCount++
-			}
-		}
-		shedPct := float64(shedCount) / math.Max(float64(issuedCount), 1) * 100
-		fmt.Printf("  %s: issued=%d granted=%d shed=%d (%.1f%%)\n",
-			tc.label, issuedCount, grantedCount, shedCount, shedPct)
 	}
 
-	fmt.Printf("\nOutput: %s\n", *outDir)
+	fmt.Printf("\nDone. TSV files in %s\n", *outDir)
+	fmt.Printf("To generate charts: python3 plot_suite.py %s\n", *outDir)
 }
